Add SQL Profiling to djblets profiling log
When profiling pages, just having the python profile isn't very helpful if 99%
of the time is spent in database queries. This change adds the ability to track
SQL queries (and where they came from). This isn't quite as awesome as it could
be regarding settings.DEBUG, but I need to think about the best way to do that.
django doesn't make it easy to replace the CursorDebugWrapper in a sane way.

Testing done:
Ran with this code and looked at profiling information.

Reviewed at http://reviews.review-board.org/r/590/