Add djblets.util.filesystem.is_exe_in_path, which will check if the specified executable is in the executable search path.
This function will do intelligent file extension determination. It expects
an application name without a file extension (such as "cvs") and will append
the ".exe" on Windows. This is an improvement over the method we used in
Review Board before where we blindly searched for "cvs" (which could in theory
exist yet not be an executable) on Windows, or "cvs.exe" (which could in
theory exist yet not be an executable) on Linux.

Reviewed at http://reviews.review-board.org/r/917/